refactor: refactor main page
