+++ src/app/features/components/main/main.scss
@@ -1,0 +1,302 @@
+@use '../../../shared/style/abstract/placeholders' as *;
+@use '../../../shared/style/abstract/constant' as *;
+@use '../../../shared/style/abstract/mixins' as *;
+
+.container {
+  max-width: $content-width;
+  margin: 0 auto;
+  padding: 1.5rem 5rem;
+
+  @include media-tablet {
+    padding: 1.5rem 2rem;
+  }
+  @include media-tablet-mobile {
+    padding: 1.5rem 1rem;
+  }
+}
+
+h1 {
+  font-size: 32px;
+  color: var(--mat-sys-primary);
+}
+
+h2 {
+  text-decoration: underline;
+  text-decoration-color: var(--mat-sys-primary);
+}
+
+.description {
+  background-color: var(--mat-sys-surface-container-low);
+
+  &__container {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    column-gap: 3rem;
+
+    @include media-tablet-mobile {
+      flex-direction: column;
+      row-gap: 1rem;
+    }
+  }
+
+  &__information {
+    display: flex;
+    flex-direction: column;
+    row-gap: 2rem;
+    width: 450px;
+
+    @include media-tablet-mobile {
+      width: 350px;
+    }
+  }
+
+  &__text {
+    font-size: 18px;
+    line-height: 1.5;
+
+    @include media-tablet {
+      font-size: 16px;
+    }
+  }
+
+  &__button {
+    border-radius: 10px;
+    padding: 1rem;
+    font-size: 16px;
+    width: 300px;
+    background-color: var(--mat-sys-primary);
+    border-color: transparent;
+    box-sizing: border-box;
+    color: var(--mat-sys-surface-container-low);
+    @extend %transition;
+
+    &:hover {
+      cursor: pointer;
+      background-color: var(--mat-sys-on-primary-container);
+    }
+  }
+
+  &__image {
+    background-image: url('../../../../assets/images/foxflow_description.png');
+    width: 530px;
+    height: 320px;
+    object-fit: cover;
+    background-repeat: round;
+
+    @include media-tablet {
+      width: 330px;
+      height: 220px;
+    }
+  }
+}
+
+.goals {
+  &__container {
+    display: flex;
+    justify-content: space-around;
+    padding: 1.5rem 0;
+    align-items: center;
+
+    @include media-tablet {
+      flex-direction: column;
+      row-gap: 1rem;
+    }
+  }
+  &__image {
+    background-image: url('../../../../assets/images/goals.png');
+    background-repeat: round;
+    width: 300px;
+    height: 300px;
+    object-fit: cover;
+
+    @include media-tablet-mobile {
+      width: 250px;
+      height: 250px;
+    }
+  }
+  &__list {
+    display: flex;
+    flex-direction: column;
+    row-gap: 1.5rem;
+    &__item {
+      display: flex;
+      justify-content: left;
+      align-items: center;
+      column-gap: 1.5rem;
+
+      @include media-tablet-mobile {
+        column-gap: 0.5rem;
+      }
+    }
+    &__img {
+      background-image: url('../../../../assets/images/checkmark.png');
+      width: 50px;
+      height: 50px;
+
+      @include media-tablet-mobile {
+        width: 30px;
+        height: 30px;
+        object-fit: cover;
+        background-repeat: round;
+      }
+    }
+  }
+}
+
+.advantages {
+  background-color: var(--mat-sys-surface-container-low);
+
+  &__container {
+    display: flex;
+    justify-content: space-around;
+    padding: 2rem 0;
+    flex-wrap: wrap;
+    row-gap: 1.5rem;
+  }
+
+  &__card {
+    padding: 1rem;
+    display: flex;
+    flex-direction: column;
+    row-gap: 1rem;
+    text-align: center;
+    width: 300px;
+    background-color: var(--mat-sys-surface);
+    border-color: var(
+      --mat-card-elevated-container-color,
+      var(--mat-sys-surface-container-low)
+    );
+    border-radius: var(
+      --mat-card-elevated-container-shape,
+      var(--mat-sys-corner-medium)
+    );
+    box-shadow: var(
+      --mat-card-elevated-container-elevation,
+      var(--mat-sys-level1)
+    );
+    @extend %transition;
+
+    &:hover {
+      box-shadow: 0 3px 7px grey;
+    }
+
+    &__image {
+      width: 100px;
+      height: 100px;
+      object-fit: cover;
+      background-repeat: no-repeat;
+      align-self: center;
+    }
+  }
+}
+.user-friendly {
+  background-image: url('../../../../assets/images/frendly_interface.png');
+}
+.tools {
+  background-image: url('../../../../assets/images/tools.png');
+}
+.workflow {
+  background-image: url('../../../../assets/images/workflow.png');
+}
+
+.members {
+  &__container {
+    display: flex;
+    justify-content: space-around;
+    flex-wrap: wrap;
+    row-gap: 1.5rem;
+    padding: 2rem 0;
+  }
+}
+
+.rs-link {
+  width: 60%;
+  border: 2px solid var(--mat-sys-primary);
+  border-radius: 10px;
+  margin: 0 auto 1rem;
+  padding: 10px;
+  @extend %transition;
+
+  @include media-tablet {
+    width: 90%;
+  }
+
+  @include media-mobile {
+    width: 100%;
+  }
+
+  a {
+    text-decoration: none;
+    color: var(--mat-sys-primary);
+    padding: 1.5rem;
+
+    &:hover {
+      color: var(--mat-sys-on-surface);
+    }
+  }
+
+  &:hover {
+    box-shadow: 0 3px 7px grey;
+    background-color: var(--mat-sys-surface-container);
+  }
+}
+
+.rs-information {
+  display: flex;
+  justify-content: center;
+  column-gap: 2rem;
+  align-items: center;
+
+  @include media-tablet {
+    column-gap: 0.5rem;
+  }
+
+  .logo-rs {
+    width: 5rem;
+    height: 5rem;
+    background-image: url('../../../../assets/images/rss-logo.svg');
+    background-repeat: no-repeat;
+
+    &__message {
+      display: flex;
+      flex-direction: column;
+      row-gap: 1rem;
+      font-weight: 600;
+
+      @include media-tablet {
+        h2 {
+          font-size: 14px;
+        }
+      }
+    }
+  }
+}
+
+.footer {
+  background-color: var(--mat-sys-surface-container);
+  color: var(--mat-sys-primary);
+
+  &__information {
+    text-align: center;
+    font-weight: 400;
+    font-size: 16px;
+  }
+}
+
+.some_class {
+  background-color: var(--mat-sys-surface-container-low);
+  border-color: var(
+    --mat-card-elevated-container-color,
+    var(--mat-sys-surface-container-low)
+  );
+  border-radius: var(
+    --mat-card-elevated-container-shape,
+    var(--mat-sys-corner-medium)
+  );
+  box-shadow: var(
+    --mat-card-elevated-container-elevation,
+    var(--mat-sys-level1)
+  );
+  text-align: justify;
+}
