--- src/app/shared/components/main/main.scss
@@ -1,149 +1,0 @@
-@use '../../style/abstract/placeholders' as *;
-@use '../../style/abstract/constant' as *;
-@use '../../style/abstract/mixins' as *;
-
-.container {
-  max-width: $content-width;
-  margin: 0 auto 10px;
-  text-align: justify;
-  background-color: var(--mat-sys-surface-container-low);
-  border-color: var(
-    --mat-card-elevated-container-color,
-    var(--mat-sys-surface-container-low)
-  );
-  border-radius: var(
-    --mat-card-elevated-container-shape,
-    var(--mat-sys-corner-medium)
-  );
-  box-shadow: var(
-    --mat-card-elevated-container-elevation,
-    var(--mat-sys-level1)
-  );
-}
-
-.main-page {
-  display: flex;
-  flex-direction: column;
-  row-gap: 1.5rem;
-  padding: 2rem;
-}
-
-h1 {
-  font-size: 32px;
-  color: var(--mat-sys-primary);
-}
-
-h2 {
-  text-decoration: underline;
-  text-decoration-color: var(--mat-sys-primary);
-}
-
-.description-text {
-  font-size: 18px;
-  line-height: 1.5;
-
-  @include media-tablet {
-    font-size: 16px;
-  }
-}
-
-.members {
-  padding: 1.5rem 0;
-  display: flex;
-  justify-content: space-around;
-  flex-wrap: wrap;
-  row-gap: 1.5rem;
-}
-
-.goals,
-.advantages,
-.description {
-  display: flex;
-  flex-direction: column;
-  row-gap: 1rem;
-
-  &__list {
-    display: flex;
-    flex-direction: column;
-    row-gap: 0.5rem;
-    margin-left: 1rem;
-    font-size: 18px;
-    line-height: 1.5;
-
-    @include media-tablet {
-      font-size: 16px;
-    }
-  }
-}
-
-.rs-link {
-  width: 60%;
-  border: 2px solid var(--mat-sys-primary);
-  border-radius: 10px;
-  margin: 0 auto 1rem;
-  padding: 10px;
-  @extend %transition;
-
-  @include media-tablet {
-    width: 90%;
-  }
-
-  @include media-mobile {
-    width: 100%;
-  }
-
-  a {
-    text-decoration: none;
-    color: var(--mat-sys-primary);
-    padding: 1.5rem;
-
-    &:hover {
-      color: var(--mat-sys-on-surface);
-    }
-  }
-
-  &:hover {
-    box-shadow: 0 3px 7px grey;
-    background-color: var(--mat-sys-surface-container);
-  }
-}
-
-.rs-information {
-  display: flex;
-  justify-content: center;
-  column-gap: 2rem;
-  align-items: center;
-
-  @include media-tablet {
-    column-gap: 0.5rem;
-  }
-
-  .logo-rs {
-    width: 5rem;
-    height: 5rem;
-    background-image: url('../../../../assets/images/rss-logo.svg');
-    background-repeat: no-repeat;
-
-    &__message {
-      display: flex;
-      flex-direction: column;
-      row-gap: 1rem;
-      font-weight: 600;
-
-      @include media-tablet {
-        h2 {
-          font-size: 14px;
-        }
-      }
-    }
-  }
-}
-
-footer {
-  padding: 1rem 0;
-  text-align: center;
-  font-weight: 400;
-  font-size: 16px;
-  background-color: var(--mat-sys-surface-container);
-  color: var(--mat-sys-primary);
-}
