+++ src/app/shared/components/member/member.html
@@ -10,6 +10,7 @@
         height="50"
         mat-card-md-image
         ngSrc="image()"
+        priority
         width="50" />
     </mat-card-title-group>
   </mat-card-header>
