+++ src/app/shared/style/abstract/_constant.scss
@@ -1,3 +1,4 @@
 $content-width: 1440px;
 $tablet-width: 768px;
 $mobile-width: 380px;
+$tablet-mobile-width: 680px;
