+++ src/app/shared/style/abstract/_mixins.scss
@@ -16,3 +16,9 @@
     @content;
   }
 }
+
+@mixin media-tablet-mobile {
+  @media (max-width: $tablet-mobile-width) {
+    @content;
+  }
+}
