+++ src/assets/i18n/en.json
@@ -39,15 +39,18 @@
   "MAIN": {
     "TITLE": "FoxFlow",
     "SUBTITLE": "About The Project",
-    "DESCRIPTION": "FoxFlow is an intuitive task management application designed to streamline your workflow and enhance productivity. FoxFlow allows you to easily create, organize, and prioritize tasks, ensuring that you stay on top of your projects. Collaborate seamlessly with team members, set deadlines, and track progress in real-time, making it the perfect tool for individuals and teams looking to optimize their task management experience",
+    "DESCRIPTION": "FoxFlow is an intuitive task management application designed to streamline your workflow and enhance productivity",
     "GOALS": "Goals",
     "GOAL_1": "Simplify task organization",
     "GOAL_2": "Enhance team collaboration",
     "GOAL_3": "Provide real-time updates and notifications",
     "ADVANTAGES": "Why Choose FoxFlow",
-    "ADVANTAGE_1": "User-Friendly Interface. FoxFlow features an intuitive design that makes it easy for users of all skill levels to navigate and manage tasks efficiently.",
-    "ADVANTAGE_2": "Collaboration Tools. The application offers robust collaboration features, allowing teams to work together seamlessly, share updates, and communicate effectively on projects.",
-    "ADVANTAGE_3": "Customizable Workflows. FoxFlow enables users to tailor their task management experience with customizable workflows, ensuring that it fits their unique project needs and preferences.",
+    "ADVANTAGE_TITLE1": "User-Friendly Interface",
+    "ADVANTAGE_TITLE2": "Collaboration Tools",
+    "ADVANTAGE_TITLE3": "Customizable Workflows",
+    "ADVANTAGE_1": "FoxFlow features an intuitive design that makes it easy for users of all skill levels to navigate and manage tasks efficiently.",
+    "ADVANTAGE_2": "The application offers robust collaboration features, allowing teams to work together seamlessly, share updates, and communicate effectively on projects.",
+    "ADVANTAGE_3": "FoxFlow enables users to tailor their task management experience with customizable workflows, ensuring that it fits their unique project needs and preferences.",
     "TEAM": "Our Team",
     "YEAR": "Created in 2025",
     "RS_MESSAGE": "FowFlow project is created as learning part of the RS School course",
