+++ src/assets/i18n/pl.json
@@ -39,15 +39,18 @@
   "MAIN": {
     "TITLE": "FoxFlow",
     "SUBTITLE": "O projekcie",
-    "DESCRIPTION": "FoxFlow to intuicyjna aplikacja do zarządzania zadaniami, zaprojektowana z myślą o usprawnieniu przepływu pracy i zwiększeniu produktywności. FoxFlow pozwala łatwo tworzyć, organizować i priorytetyzować zadania, zapewniając pełną kontrolę nad projektami. Bezproblemowa współpraca z członkami zespołu, ustalanie terminów i śledzenie postępów w czasie rzeczywistym – to idealne narzędzie dla osób i zespołów, które chcą zoptymalizować zarządzanie zadaniami.",
+    "DESCRIPTION": "FoxFlow to intuicyjna aplikacja do zarządzania zadaniami, która usprawnia twój przepływ pracy i zwiększa produktywność",
     "GOALS": "Cele",
     "GOAL_1": "Uprościć organizację zadań",
     "GOAL_2": "Ulepszenie współpracy zespołowej",
     "GOAL_3": "Podaj aktualizacje i powiadomienia w czasie rzeczywistym",
     "ADVANTAGES": "Dlaczego warto wybrać FoxFlow",
-    "ADVANTAGE_1": "Przyjazny dla użytkownika interfejs. FoxFlow charakteryzuje się intuicyjnym interfejsem, który ułatwia użytkownikom o różnym poziomie umiejętności nawigację i efektywne zarządzanie zadaniami.",
-    "ADVANTAGE_2": "Narzędzia do współpracy. Aplikacja oferuje rozbudowane funkcje współpracy, umożliwiając zespołom bezproblemową współpracę, udostępnianie aktualizacji i efektywną komunikację w ramach projektów.",
-    "ADVANTAGE_3": "Konfigurowalne przepływy pracy. FoxFlow umożliwia użytkownikom dostosowanie sposobu zarządzania zadaniami dzięki konfigurowalnym przepływom pracy, które odpowiadają ich unikalnym potrzebom i preferencjom projektu.",
+    "ADVANTAGE_TITLE1": "Przyjazny dla użytkownika interfejs",
+    "ADVANTAGE_TITLE2": "Narzędzia do współpracy",
+    "ADVANTAGE_TITLE3": "Konfigurowalne przepływy pracy",
+    "ADVANTAGE_1": "FoxFlow charakteryzuje się intuicyjnym interfejsem, który ułatwia użytkownikom o różnym poziomie umiejętności nawigację i efektywne zarządzanie zadaniami.",
+    "ADVANTAGE_2": "Aplikacja oferuje rozbudowane funkcje współpracy, umożliwiając zespołom bezproblemową współpracę, udostępnianie aktualizacji i efektywną komunikację w ramach projektów.",
+    "ADVANTAGE_3": "FoxFlow umożliwia użytkownikom dostosowanie sposobu zarządzania zadaniami dzięki konfigurowalnym przepływom pracy, które odpowiadają ich unikalnym potrzebom i preferencjom projektu.",
     "TEAM": "Nasz Zespół",
     "YEAR": "Utworzono w 2025 roku",
     "RS_MESSAGE": "Projekt FowFlow powstał jako część kursu RS School",
